Add 'IL GARS Annual Report FY2014.pdf' csv files
The annual GARS report is a pdf file and thus the tables are difficult
to read programmatically.  This adds a non-exhaustive directory of csv
files to the repository.  The csv files were created with the ease of
importing with python or R in mind.  If errors occur due to formatting,
it is always possible to get the info from the original pdf.  Lastly,
there is a considerable amount of repetetive informaiton, but each csv
file contains at least one piece of exclusive information.  Future work
can be put into further organizing the informaiton.